backend/morningai-api/requirements.txt
# FastAPI 核心依賴
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 數據庫相關
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis 快取
redis==5.0.1
aioredis==2.0.1

# 認證與安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP 客戶端
httpx==0.25.2
aiohttp==3.9.1

# 外部 API 整合
openai==1.3.7
python-telegram-bot==20.7
line-bot-sdk==3.5.0

# 工具庫
python-dotenv==1.0.0
email-validator==2.1.0
celery==5.3.4
flower==2.0.1

# 監控與日誌
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# 開發工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
bandit==1.7.5

# CORS 支援
fastapi-cors==0.0.6

# 限流
slowapi==0.1.9

# 文件上傳
python-magic==0.4.27
pillow==10.1.0

# 時間處理
python-dateutil==2.8.2
pytz==2023.3

# 配置管理
dynaconf==3.2.4

# 測試工具
factory-boy==3.3.0
faker==20.1.0
